added another attributes file
