+++ test/sample_data/types.ts
@@ -20,6 +20,9 @@ import ManualAssignment from "../../src/models/manual/assignment";
 import Membership from "../../src/models/membership";
 import MembershipRequest from "../../src/models/membership_request";
 
+/**
+ * Set of types so we can iterate over them in tests
+ */
 const types: { [i: string]: new () => unknown } = {
     business: Business,
     membership: Membership,
